fix: vestings latest height
