+++ site/src/pages/index.js
@@ -1,22 +1,10 @@
-import React, { useEffect } from "react";
+import React from "react";
 import HomeLayout from "../components/layout/home";
 import Explore from "../components/home/explore";
 import Sections from "../components/home/sections";
-import {
-  connect,
-  disconnect,
-  unSubscribeHomepageInfo,
-} from "../services/websocket";
-import { useDispatch } from "react-redux";
-import {
-  setLatestBlocks,
-  setLatestSignedTransfers,
-} from "../store/reducers/socketSlice";
 import styled from "styled-components";
 import Overview from "../components/home/overview";
 import { m_t } from "../styles/tailwindcss";
-import { clearNftList } from "../store/reducers/nftSlice";
-import { clearAssetList } from "../store/reducers/assetSlice";
 
 const SectionsWrapper = styled.div`
   ${m_t(16)};
@@ -27,20 +15,6 @@ const OverviewWrapper = styled.div`
 `;
 
 function Home() {
-  const dispatch = useDispatch();
-  useEffect(() => {
-    connect();
-
-    return () => {
-      unSubscribeHomepageInfo();
-      disconnect();
-      dispatch(setLatestSignedTransfers([]));
-      dispatch(setLatestBlocks([]));
-      dispatch(clearNftList());
-      dispatch(clearAssetList());
-    };
-  }, [dispatch]);
-
   return (
     <HomeLayout>
       <Explore />
